fix: resolve all build errors (SID C4146, GUI C2509, linker LNK2001)
- sid_waveform_tables.cpp: cast to int32_t before negation to fix C4146
- mos6581.h: move render_* declarations outside #ifdef IMGUI_VERSION guard
- tests/CMakeLists.txt: add GUI stub sources to mos6510_lib,
  fam65xx_processor_tests_runner, and sid_test_runner targets

+++ src/chip/sound/sid_waveform_tables.cpp
@@ -466,7 +466,7 @@ void init_waveform_tables() {
         model_wave[0][4][i] = model_wave[1][4][i] = 0xfff;
 
         // Triangle: ((accumulator ^ -!!msb) >> 11) & 0xffe
-        tri[i] = ((accumulator ^ -(uint32_t)!!msb) >> 11) & 0xffe;
+        tri[i] = ((accumulator ^ (uint32_t)(-(int32_t)!!msb)) >> 11) & 0xffe;
         model_wave[0][1][i] = model_wave[1][1][i] = tri[i];
 
         // Sawtooth: accumulator >> 12
